Refactor authentication logic into multiple files for increased understandability
This change also modifies some logic to support one wallet at a time, instead of multi-wallets

The User provider had a large amount of authentication and partnership related logic. This logic was unbundled and placed into different files under the wallets-and-connections directory. These files include disconnect-wallet.ts, get-unauthenticated-wallets.ts, get-authenticated-wallets.ts and authenticate-and-load-partnerships.ts. This change should make it easier to understand the logic behind each authentication action and improve maintainability of the codebase.
